% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suelos_CIREN.R
\docType{data}
\name{Suelos_CIREN}
\alias{Suelos_CIREN}
\alias{suelo_ciren_r05}
\alias{suelo_ciren_r06}
\alias{suelo_ciren_r07}
\alias{suelo_ciren_r08}
\alias{suelo_ciren_r09}
\alias{suelo_ciren_r10_cont}
\alias{suelo_ciren_r10_chil}
\alias{suelo_ciren_r13}
\alias{suelo_ciren_r14}
\alias{suelo_ciren_r16}
\title{Suelos CIREN}
\format{
\describe{
\item{DESCEROS}{Descripción de la erosión del suelo}
\item{DESCPEND}{Descripción 1 de la pendiente del suelo}
\item{DESCPROF}{Descripción 2 de la profundidad del suelo}
\item{DESCSUCA}{Descripción 1 de la sub-capacidad de uso del suelo}
\item{MSLINK}{Identificador único del elemento gráfico}
\item{NOMBSERI}{Nombre de la serie}
\item{PORCPOLI}{Porcentaje de suelo en la parametría indicada}
\item{SIMBSERI}{Símbolo de la serie de suelo}
\item{TEXTCAUS}{Descripción de la capacidad de uso del suelo}
\item{TEXTPEND}{Descripción 2 de la pendiente del suelo}
\item{TEXTPROF}{Descripción 2 de la profundidad del suelo}
\item{TEXTSUCA}{Descripción 2 de la sub-capacidad de uso del suelo}
}

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6494 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10975 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13261 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5979 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5415 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2980 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4387 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2079 rows and 13 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13548 rows and 13 columns.
}
\source{
\url{https://www.ciren.cl/}
}
\usage{
suelo_ciren_r05

suelo_ciren_r06

suelo_ciren_r07

suelo_ciren_r08

suelo_ciren_r09

suelo_ciren_r10_cont

suelo_ciren_r10_chil

suelo_ciren_r13

suelo_ciren_r14

suelo_ciren_r16
}
\description{
Datos geoespaciales de suelos en Chile desde Valparaíso a Los Lagos.
Estos se presentan en coordenadas WGS84 (EPSG:4326).
}
\keyword{datasets}
